requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
# 用于提供静态文件服务
aiofiles==23.2.1

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
# psycopg2-binary==2.9.9  # PostgreSQL adapter - uncomment if using PostgreSQL
# sqlite3  # Built-in with Python

# Authentication and Security
python-jose==3.3.0
cryptography==3.4.8
passlib==1.7.4
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.11.7
pydantic-settings==2.10.1

# Asynchronous task queue
celery==5.3.6
redis==5.0.1

# AI and RAG
openai==1.3.6
langchain==0.0.340
chromadb==0.4.18
tiktoken==0.5.2

# Utilities
python-dotenv==1.1.1
requests==2.31.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional dependencies (uncomment if needed)
# PostgreSQL support:
# psycopg2-binary==2.9.9